refactor(preset-mini|preset-wind): move functions around - move `colorResolver` from color to utils - move `parseColorUtil` from color to utils as `parseColor` - adjust imports - split `auto` handler & adjust usage
