+++ packages/preset-mini/src/rules/svg.ts
@@ -1,6 +1,5 @@
 import type { Rule } from '@unocss/core'
-import { handler as h } from '../utils'
-import { colorResolver } from './color'
+import { colorResolver, handler as h } from '../utils'
 
 export const svgUtilities: Rule[] = [
   // fills
