+++ packages/preset-mini/src/rules/typography.ts
@@ -1,8 +1,7 @@
 import type { Rule } from '@unocss/core'
 import { toArray } from '@unocss/core'
 import type { Theme } from '../theme'
-import { handler as h } from '../utils'
-import { colorResolver } from './color'
+import { colorResolver, handler as h } from '../utils'
 
 const weightMap: Record<string, string> = {
   thin: '100',
@@ -30,7 +29,7 @@ export const fonts: Rule<Theme>[] = [
 
   // size
   [/^text-(.+)$/, ([, s = 'base'], { theme }) => {
-    const size = h.bracket.rem(s)
+    const size = h.bracket.auto.rem(s)
     if (size)
       return { 'font-size': size }
 
@@ -44,7 +43,7 @@ export const fonts: Rule<Theme>[] = [
     }
   }],
   [/^text-size-(.+)$/, ([, s]) => {
-    const raw = h.bracket.rem(s)
+    const raw = h.bracket.auto.rem(s)
     if (raw)
       return { 'font-size': raw }
   }],
@@ -58,21 +57,21 @@ export const fonts: Rule<Theme>[] = [
 
   // leadings
   [/^(?:leading|lh)-([^-]+)$/, ([, s], { theme }) => {
-    const v = theme.lineHeight?.[s] || h.bracket.rem(s)
+    const v = theme.lineHeight?.[s] || h.bracket.auto.rem(s)
     if (v !== null)
       return { 'line-height': v }
   }],
 
   // tracking
   [/^tracking-([^-]+)$/, ([, s], { theme }) => {
-    const v = theme.letterSpacing?.[s] || h.bracket.rem(s)
+    const v = theme.letterSpacing?.[s] || h.bracket.auto.rem(s)
     if (v !== null)
       return { 'letter-spacing': v }
   }],
 
   // word-spacing
   [/^word-spacing-([^-]+)$/, ([, s], { theme }) => {
-    const v = theme.wordSpacing?.[s] || h.bracket.rem(s)
+    const v = theme.wordSpacing?.[s] || h.bracket.auto.rem(s)
     if (v !== null)
       return { 'word-spacing': v }
   }],
@@ -94,7 +93,7 @@ export const tabSizes: Rule<Theme>[] = [
 
 export const textIndents: Rule<Theme>[] = [
   [/^indent(?:-(.+))?$/, ([, s], { theme }) => {
-    const v = theme.textIndent?.[s || 'DEFAULT'] || h.bracket.cssvar.fraction.rem(s)
+    const v = theme.textIndent?.[s || 'DEFAULT'] || h.bracket.cssvar.fraction.auto.rem(s)
     if (v != null)
       return { 'text-indent': v }
   }],
