+++ packages/preset-mini/src/utils/utilities.ts
@@ -1,4 +1,6 @@
-import type { CSSEntries } from '@unocss/core'
+import type { CSSEntries, RuleContext } from '@unocss/core'
+import { hex2rgba } from '@unocss/core'
+import type { Theme } from '../theme'
 import { handler as h } from './handlers'
 import { directionMap } from './mappings'
 
@@ -7,7 +9,105 @@ export function capitalize<T extends string>(str: T) {
 }
 
 export const directionSize = (prefix: string) => ([_, direction, size]: string[]): CSSEntries | undefined => {
-  const v = h.bracket.rem.fraction.cssvar(size)
-  if (v)
-    return directionMap[direction].map(i => [prefix + i, v])
+  const v = h.bracket.auto.rem.fraction.cssvar(size)
+  if (v !== undefined)
+    return directionMap[direction].map(i => [`${prefix}${i}`, v])
+}
+
+const getThemeColor = (theme: Theme, colors: string[]) =>
+  theme.colors?.[
+    colors.join('-').replace(/(-[a-z])/g, n => n.slice(1).toUpperCase())
+  ]
+
+export const parseColor = (body: string, theme: Theme) => {
+  const [main, opacity] = body.split(/(?:\/|:)/)
+  const colors = main
+    .replace(/([a-z])([0-9])/g, '$1-$2')
+    .split(/-/g)
+  const [name] = colors
+
+  if (!name)
+    return
+
+  let color: string | undefined
+  const bracket = h.bracket(main)
+  const bracketOrMain = bracket || main
+
+  if (bracketOrMain.startsWith('#'))
+    color = bracketOrMain.slice(1)
+  if (bracketOrMain.startsWith('hex-'))
+    color = bracketOrMain.slice(4)
+
+  color = color || bracket
+
+  let no = 'DEFAULT'
+  if (!color) {
+    let colorData
+    const [scale] = colors.slice(-1)
+    if (scale.match(/^\d+$/)) {
+      no = scale
+      colorData = getThemeColor(theme, colors.slice(0, -1))
+    }
+    else {
+      colorData = getThemeColor(theme, colors)
+      if (!colorData) {
+        [, no = no] = colors
+        colorData = getThemeColor(theme, [name])
+      }
+    }
+
+    if (typeof colorData === 'string')
+      color = colorData
+    else if (no && colorData)
+      color = colorData[no]
+  }
+
+  return {
+    opacity,
+    name,
+    no,
+    color,
+    rgba: hex2rgba(color),
+  }
+}
+
+export const colorResolver = (attribute: string, varName: string) => ([, body]: string[], { theme }: RuleContext<Theme>) => {
+  const data = parseColor(body, theme)
+
+  if (!data)
+    return
+
+  const { opacity, color, rgba } = data
+
+  if (!color)
+    return
+
+  const a = opacity
+    ? opacity[0] === '['
+      ? h.bracket.percent(opacity)!
+      : (parseFloat(opacity) / 100)
+    : rgba?.[3]
+
+  if (rgba) {
+    if (a != null && !Number.isNaN(a)) {
+      // @ts-expect-error
+      rgba[3] = typeof a === 'string' && !a.includes('%')
+        ? parseFloat(a)
+        : a
+      return {
+        [attribute]: `rgba(${rgba.join(',')})`,
+      }
+    }
+    else {
+      return {
+        [`--un-${varName}-opacity`]: 1,
+        [attribute]: `rgba(${rgba.slice(0, 3).join(',')},var(--un-${varName}-opacity))`,
+      }
+    }
+  }
+  else {
+    return {
+      [attribute]: color.replace('%alpha', `${a || 1}`),
+    }
+  }
 }
