+++ packages/preset-wind/src/rules/behaviors.ts
@@ -1,6 +1,5 @@
 import type { Rule } from '@unocss/core'
-import { colorResolver } from '@unocss/preset-mini/rules'
-import { handler as h } from '@unocss/preset-mini/utils'
+import { colorResolver, handler as h } from '@unocss/preset-mini/utils'
 
 const listStyleProps = ['none', 'disc', 'circle', 'square', 'decimal', 'zero-decimal', 'greek', 'roman', 'upper-roman', 'alpha', 'upper-alpha']
