+++ packages/preset-wind/src/rules/divide.ts
@@ -1,6 +1,5 @@
 import type { CSSEntries, Rule } from '@unocss/core'
-import { directionMap, handler as h } from '@unocss/preset-mini/utils'
-import { colorResolver } from '@unocss/preset-mini/rules'
+import { colorResolver, directionMap, handler as h } from '@unocss/preset-mini/utils'
 
 export const divideColors: Rule[] = [
   [/^divide-(.+)$/, colorResolver('border-color', 'divide')],
